requirements.txt
opencv-python
albumentations
pillow
pandas
sklearn
tensorboard_logger
pycm
tqdm
scipy
torchnet
matplotlib
